Refactored FormatDatabase in ROS1 and ROS2 versions
