+++ src/roswire/ros2/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from .format import ROS2FormatDatabase
 from .launch import ROS2LaunchManager
 from .node_manager import ROS2NodeManager
 from .package import ROS2PackageDatabase
